+++ posydon/grids/psygrid.py
@@ -221,6 +221,7 @@ from posydon.utils.common_functions import (
 )
 from posydon.utils.configfile import ConfigFile
 from posydon.utils.gridutils import (
+    _get_grid_column,
     add_field,
     fix_He_core,
     join_lists,
